chore(core-terms-and-conditions): do work
